requirements.txt
# Core
streamlit==1.40.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.32.0

# Google Gemini
google-generativeai==0.8.3

# LangChain Ecosystem (fully compatible)
langchain==0.3.7
langchain-core==0.3.17
langchain-google-genai==2.0.5
langchain-community==0.3.7

# Vector DB
pinecone>=3.0.0
sentence-transformers==3.3.1

# MongoDB Drivers
pymongo==4.9.1
motor==3.6.0

# Email Processing
email-validator==2.2.0
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
tenacity==9.0.0
requests==2.32.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
